Cleaning up error outputs for dnx and dnu to address #864 by hiding stack traces for non-user errors
