+++ misc/BootstrapperSolution/global.json
@@ -1,0 +1,3 @@
+﻿{
+  "projects": [ "src" ]
+}
