+++ misc/BootstrapperSolution/src/TesterProgram/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace TesterProgram
+{
+    public class Program
+    {
+        public void Main(string[] args)
+        {
+            throw new Exception("foo");
+        }
+    }
+}
