+++ test/Bootstrapper.FunctionalTests/project.json
@@ -2,6 +2,7 @@
     "dependencies": {
         "Microsoft.Dnx.CommonTestUtils": "1.0.0-*",
         "Microsoft.Dnx.Runtime.Sources": "1.0.0-*",
+        "Microsoft.Dnx.Testing.CommonUtils": "1.0.0-*",
         "Microsoft.AspNet.Testing": "1.0.0-*",
         "xunit.runner.aspnet": "2.0.0-aspnet-*"
     },
